fix(validation): make live-source analysis honest and resilient
- compare explicit claimed coordinates instead of silently ignoring them\n- keep single-source cases at INSUFFICIENT_SOURCES\n- prevent a provider from corroborating its own report or being queried twice\n- scope L1 cache entries by ICAO and claim source\n- exclude generic fused ADSB tracks from external source validation\n- enrich API fallback records and add adsb.lol ingestion fallback for OpenSky 429 responses\n- report Redis and PostgreSQL failures through health checks\n- convert NumPy anomaly scores to JSON-safe native floats\n- replace assertion-free integration output with deterministic synthetic timing assertions\n- add regressions for validation, fallback, health, ingestion, caching, and serialization

+++ anomaly/detector.py
@@ -210,7 +210,7 @@ class AircraftBaseline:
         sigma = arr.std()
         if sigma < 1e-6:
             return 0.0
-        return abs((value - mu) / sigma)
+        return float(abs((value - mu) / sigma))
 
 
 class StatisticalDetector:
